refactored some unit-tests structure
