+++ dataprofiler/tests/data_readers/test_data_utils.py
@@ -2,7 +2,6 @@ import os
 import unittest
 from itertools import islice
 
-from dataprofiler import dp_logging
 from dataprofiler.data_readers import data_utils
 
 test_root_path = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
@@ -238,20 +237,3 @@ class TestDataReadingWriting(unittest.TestCase):
                 chunk_size_bytes=f["chunk_size_bytes"],
             )
             self.assertEqual(expected, output_str, f["path"])
-
-    def test_is_s3_uri_failure_logger_check(self):
-        invalid_path = "invalid_path"
-
-        logger = dp_logging.get_child_logger(__name__)
-
-        with self.assertLogs(logger, level="DEBUG") as log_context:
-            # Call the function with the invalid path
-            is_s3 = data_utils.S3Helper.is_s3_uri(invalid_path, logger)
-
-            # Assert that the function returns False (invalid path)
-            self.assertFalse(is_s3)
-
-            # Assert that the log message is generated and logged
-            self.assertIn(
-                f"'{invalid_path}' is not a valid S3 URI", log_context.output[0]
-            )
